Moved atoms to atom folder
